+++ spec/hydra_attribute/active_record_spec.rb
@@ -1,6 +1,123 @@
 require 'spec_helper'
 
 describe HydraAttribute::ActiveRecord do
+  describe '.has_many' do
+    context 'dependent: :destroy' do
+      before do
+        Room.hydra_attributes.create(name: 'length', backend_type: 'integer')
+        Room.hydra_attributes.create(name: 'width',  backend_type: 'integer')
+      end
+
+      context 'find' do
+        let!(:flat)  { Flat.create! }
+        let!(:room1) { Room.create!(flat_id: flat.id, square: 40, length: 5, width: 8) }
+        let!(:room2) { Room.create!(flat_id: flat.id, square: 45, length: 5, width: 9) }
+
+        it 'returns hydra associations with loaded hydra attributes' do
+          rooms = flat.rooms
+          expect(rooms).to match_array([room1, room2])
+
+          attributes = [rooms[0].attributes.except('updated_at', 'created_at')]
+          attributes << rooms[1].attributes.except('updated_at', 'created_at')
+          expect(attributes).to match_array([
+            room1.attributes.except('updated_at', 'created_at'),
+            room2.attributes.except('updated_at', 'created_at')
+          ])
+        end
+
+        it 'returns hydra associations filtered by hydra attribute' do
+          rooms = flat.rooms.where(length: 5, width: 9)
+          expect(rooms).to match_array([room2])
+          expect(rooms.first.length).to eq(5)
+          expect(rooms.first.width).to  eq(9)
+          expect(rooms.first.square).to eq(45)
+        end
+
+        it 'returns hydra associations with filter by static attribute' do
+          rooms = flat.rooms.where(square: 40)
+          expect(rooms).to match_array([room1])
+          expect(rooms.first.length).to eq(5)
+          expect(rooms.first.width).to  eq(8)
+          expect(rooms.first.square).to eq(40)
+        end
+      end
+
+      context 'build' do
+        let(:flat) { Flat.new }
+
+        it 'returns new hydra association with hydra attributes' do
+          room = flat.rooms.build(square: 20, length: 2, width: 10)
+          expect(room.square).to eq(20)
+          expect(room.length).to eq(2)
+          expect(room.width).to  eq(10)
+        end
+      end
+
+      context 'create' do
+        let(:flat) { Flat.create! }
+
+        it 'saves hydra associations with hydra attributes' do
+          room = flat.rooms.create!(square: 20, length: 2, width: 10)
+          expect(room.square).to eq(20)
+          expect(room.length).to eq(2)
+          expect(room.width).to  eq(10)
+        end
+      end
+
+      context 'destroy' do
+        let!(:flat)  { Flat.create! }
+        let!(:room1) { Room.create!(flat_id: flat.id, square: 40, length: 5, width: 8) }
+        let!(:room2) { Room.create!(flat_id: flat.id, square: 45, length: 5, width: 9) }
+
+        it 'removes hydra associations and theirs hydra values' do
+          flat.destroy
+          expect(Room.count).to be(0)
+
+          query = HydraAttribute::HydraValue.arel_tables[Room.table_name]['integer'].project(Arel.star.count)
+          count = ActiveRecord::Base.connection.select_value(query).to_i
+          expect(count).to be(0)
+        end
+      end
+    end
+  end
+
+  describe '.belongs_to' do
+    before do
+      Flat.hydra_attributes.create(name: 'floor', backend_type: 'integer')
+    end
+
+    context 'find' do
+      let!(:flat) { Flat.create!(number: 100, floor: 5) }
+      let!(:room) { Room.create!(flat_id: flat.id) }
+
+      it 'returns hydra association' do
+        expect(room.flat).to eq(flat)
+        expect(room.flat.number).to be(100)
+        expect(room.flat.floor).to  be(5)
+      end
+    end
+
+    context 'build' do
+      let(:room) { Room.new }
+
+      it 'returns hydra association' do
+        flat = room.build_flat(number: 100, floor: 3)
+        expect(flat.number).to be(100)
+        expect(flat.floor).to  be(3)
+      end
+    end
+
+    context 'create' do
+      let(:room) { Room.create! }
+
+      it 'returns hydra association' do
+        flat = room.create_flat!(number: 100, floor: 3)
+        expect(flat.number).to be(100)
+        expect(flat.floor).to  be(3)
+      end
+    end
+  end
+
   describe '.inspect' do
     before do
       HydraAttribute::HydraAttribute.create(entity_type: 'Product', name: 'code',  backend_type: 'string', white_list: true)
@@ -415,6 +532,7 @@ describe HydraAttribute::ActiveRecord do
 
     it 'should select only specific attributes and return models which have all these attributes in the attribute set' do
       relation = Product.select([:name, :code])
+
       relation.map(&:name).should =~ %w[name1 name2 name4]
       relation.map(&:code).should =~ %w[code1 code2 code4]